Finished 3rd and final acceleration video
Same issue as before. Code is already finished in the starter files. As for the issue at the end of the code a fix would be to instead apply a force in the forward direction of the turret transform at the start of the bullets life so that the bullet can be rotated without affecting its trajectory.